Added a new parser for .3DS files. Also fixed some bugs in the abstract parser and texture atlas.
git-svn-id: http://min3d.googlecode.com/svn/trunk@44 470a5c22-fb99-cafe-615f-f07c7e978966

+++ src/min3d/parser/ParseObjectFace.java
@@ -1,9 +1,9 @@
 package min3d.parser;
 
 public class ParseObjectFace {
-	public short[] v;
-	public short[] uv;
-	public short[] n;
+	public int[] v;
+	public int[] uv;
+	public int[] n;
 	public int faceLength;
 	public boolean hasuv;
 	public boolean hasn;
